requirements.txt
Celery==4.0.2
Adafruit-DHT==1.3.2
cffi==1.5.2
cryptography==1.2.3
Django==1.8.3
django-filter==0.10.0
djangorestframework==3.1.3
enum34==1.1.2
gTTS==1.1.4
gTTS-token==1.0.1
idna==2.0
ipaddress==1.0.16
Markdown==2.6.2
mcpi==0.1.1
mutagen==1.20
MySQL-python==1.2.3
ndg-httpsclient==0.4.0
numpy==1.6.2
picamera==1.10
pifacecommon==4.1.2
pifacedigitalio==3.0.4
pigpio==1.20
psutil==3.2.1
pyasn1==0.1.9
PyAudio==0.2.4
pycparser==2.14
pygame===1.9.1release
pygobject==3.8.2
pyOpenSSL==0.15.1
pyserial==2.5
requests==2.7.0
RPi.GPIO==0.6.2
six==1.10.0
SpeechRecognition==1.4.0
virtualenv==13.1.0
wolframalpha==1.2
